+++ cmd/main.go
@@ -3,11 +3,14 @@ package main
 import (
 	"fmt"
 	"os"
+	"strings"
 
 	"github.com/urfave/cli"
 
+	"github.com/manifoldco/go-manifold"
 	"github.com/manifoldco/manifold-cli/config"
 	"github.com/manifoldco/manifold-cli/plugins"
+	"github.com/manifoldco/manifold-cli/prompts"
 )
 
 var cmds []cli.Command
@@ -66,4 +69,65 @@ var helpCommand = cli.Command{
 		cli.ShowAppHelp(c)
 		return nil
 	},
+}
+
+// generateName makes a title lowercase and replace spaces with dashes
+func generateName(title string) manifold.Label {
+	name := strings.Replace(strings.ToLower(title), " ", "-", -1)
+	return manifold.Label(name)
+}
+
+// generateTitle makes a name capitalized and replace dashes with spaaces
+func generateTitle(name string) manifold.Name {
+	title := strings.Title(strings.Replace(name, "-", " ", -1))
+	return manifold.Name(title)
+}
+
+// promptNameAndTitle encapsulates the logic for accepting a name as the first
+// positional argument (optionally), and a title flag (optionally)
+// returning generated values accepted by the user
+func promptNameAndTitle(ctx *cli.Context, objectName string, shouldInferTitle, allowEmpty bool) (string, string, error) {
+	// The user may supply a name value as the first positional arg
+	argName, err := optionalArgName(ctx, 0, "name")
+	if err != nil {
+		return "", "", err
+	}
+
+	// The user may supply a title value from a flag, validate it
+	flagTitle := ctx.String("title")
+
+	// Create the title based on the name argument
+	return createNameAndTitle(ctx, objectName, argName, flagTitle, shouldInferTitle, true, allowEmpty)
+}
+
+func createNameAndTitle(ctx *cli.Context, objectName, argName, flagTitle string, shouldInferTitle, shouldAccept, allowEmpty bool) (string, string, error) {
+	var name, title string
+
+	// If no name value is supplied, prompt for it
+	// otherwise validate the given value
+	shouldAcceptName := shouldAccept && argName != ""
+	nameValue, err := prompts.Name(objectName, argName, shouldAcceptName, allowEmpty)
+	if err != nil {
+		return name, title, err
+	}
+	name = nameValue
+
+	// We will automatically validate/accept a title given as flag
+	shouldAcceptTitle := shouldAccept && flagTitle != ""
+	// If we shouldn't infer the title, do not automatically accept a title value
+	if !shouldInferTitle {
+		shouldAcceptTitle = false
+	}
+	defaultTitle := flagTitle
+	if flagTitle == "" && shouldInferTitle {
+		// If no flag is present, we will infer the title from the validated name
+		defaultTitle = string(generateTitle(name))
+	}
+	titleValue, err := prompts.Title(objectName, defaultTitle, shouldAcceptTitle, allowEmpty)
+	if err != nil {
+		return name, title, err
+	}
+	title = titleValue
+
+	return name, title, nil
 }